progresando en automatizacion de notificaciones de aplicaciones a proyectos
